Move SourceInfo into source.hpp/cpp
