+++ src/mbgl/map/tile_data.cpp
@@ -1,7 +1,7 @@
 #include <mbgl/map/tile_data.hpp>
 #include <mbgl/map/map.hpp>
 #include <mbgl/map/environment.hpp>
-#include <mbgl/style/style_source.hpp>
+#include <mbgl/map/source.hpp>
 
 #include <mbgl/util/token.hpp>
 #include <mbgl/util/string.hpp>
